% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{b}
\alias{b}
\title{Service continuous density distribution}
\usage{
b(t)
}
\arguments{
\item{t}{time value}
}
\value{
density function value (double) given \code{t}
}
\description{
Service continuous density distribution
}
