% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{densityOfSojournTimeAtState_i}
\alias{densityOfSojournTimeAtState_i}
\title{The density of the sojourn time in state i with probability that}
\usage{
densityOfSojournTimeAtState_i(i, j, t, dt, m = c(0.2, 0.3), mMax = 14)
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{t}{time value}

\item{dt}{time increment}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{mMax}{max number of addends in sums}
}
\value{
double
}
\description{
The density of the sojourn time in state i with probability that
}
\examples{
densityOfSojournTimeAtState_i(1, 0, 10, 1, m=c(1, 2), mMax=5)
}
