% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{EN}
\alias{EN}
\title{Expectation of number of arriving claims depending on i and j}
\usage{
EN(i, j, t, m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{t}{time value}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
double
}
\description{
Expectation of number of arriving claims depending on i and j
}
\examples{
EN(1, 1, 2)
}
