% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{ENU}
\alias{ENU}
\title{Expectation of number of arriving claims}
\usage{
ENU(i, t)
}
\arguments{
\item{i}{MC i-th state}

\item{t}{time value}
}
\value{
double
}
\description{
Expectation of number of arriving claims
}
\examples{
ENU(1, 3)
}
