% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{finalStateProbability}
\alias{finalStateProbability}
\title{Probability of the final state}
\usage{
finalStateProbability(i, j, t, m = c(0.2, 0.3))
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{t}{time value}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}
}
\value{
double
}
\description{
Probability of the final state
}
\examples{
finalStateProbability(0, 1, 10)
}
