% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{flowIntensityMean}
\alias{flowIntensityMean}
\title{The mean intensity of the arrived flow}
\usage{
flowIntensityMean(lambda = c(1, 2))
}
\arguments{
\item{lambda}{Poisson flow intensity vector}
}
\description{
The mean intensity of the arrived flow
}
