% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{h}
\alias{h}
\title{Density of empty time for initial state i jointly with probability of final state j}
\usage{
h(i, j, t, m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{t}{time value}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
double
}
\description{
Density of empty time for initial state i jointly with probability of final state j
}
\examples{
h(1, 1, 2, m = c(2.5, 0.2))
}
