% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{loadCoefficient}
\alias{loadCoefficient}
\title{Load coefficient}
\usage{
loadCoefficient(m, lambda)
}
\arguments{
\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
load coefficient value (double) of the arriving flow
}
\description{
Load coefficient
}
\examples{
loadCoefficient(m = c(0.2, 0.3), lambda = c(1,2))
}
