% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{meanSojournTimeWithFSP}
\alias{meanSojournTimeWithFSP}
\title{Mean sojourn time in the initial state i jointly with final probability of state j}
\usage{
meanSojournTimeWithFSP(i, j, t, m = c(0.2, 0.3))
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{t}{time value}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}
}
\value{
double
}
\description{
Mean sojourn time in the initial state i jointly with final probability of state j
}
\examples{
meanSojournTimeWithFSP(1, 0, 3)
}
