% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{meanSoujournTime}
\alias{meanSoujournTime}
\title{Mean sojourn time in the initial state i (without final probability of state j)}
\usage{
meanSoujournTime(i, t)
}
\arguments{
\item{i}{MC i-th state}

\item{t}{time value}
}
\value{
double
}
\description{
Mean sojourn time in the initial state i (without final probability of state j)
}
\examples{
meanSoujournTime(0, 10)
}
