% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{meanTimeEmptyFixed}
\alias{meanTimeEmptyFixed}
\title{Mean time of empty period in fixed state i}
\usage{
meanTimeEmptyFixed(i)
}
\arguments{
\item{i}{MC i-th state}
}
\value{
complex
}
\description{
Mean time of empty period in fixed state i
}
