% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{meanTimeOfBusyPeriodETW}
\alias{meanTimeOfBusyPeriodETW}
\title{Mean time of busy period}
\usage{
meanTimeOfBusyPeriodETW(m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector
description}
}
\value{
complex
}
\description{
Mean time of busy period
}
