% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{meanTimeOfBusyPeriodEW}
\alias{meanTimeOfBusyPeriodEW}
\title{Mean time of busy period multiplied by load coefficient}
\usage{
meanTimeOfBusyPeriodEW(m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
complex
}
\description{
Mean time of busy period multiplied by load coefficient
}
