% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{meanTimeOfEmptyPeriod}
\alias{meanTimeOfEmptyPeriod}
\title{Mean time of empty period given the stationary probability}
\usage{
meanTimeOfEmptyPeriod()
}
\value{
complex
}
\description{
Mean time of empty period given the stationary probability
}
