% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{MET}
\alias{MET}
\title{Mean idle time if initial state i}
\usage{
MET(i, m = c(0.2, 0.3), lambda = c(1, 2), tmax = 12)
}
\arguments{
\item{i}{MC i-th state}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}

\item{tmax}{upper integration limit}
}
\value{
double
}
\description{
Mean idle time if initial state i
}
\examples{
MET(1)
}
