% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{MST}
\alias{MST}
\title{Mean empty time sojourn time in the initial state i during the empty period}
\usage{
MST(i, m = c(0.2, 0.3), lambda = c(1, 2), tmax = 12)
}
\arguments{
\item{i}{MC i-th state}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}

\item{tmax}{upper integration limit}
}
\value{
double
}
\description{
Mean empty time sojourn time in the initial state i during the empty period
}
\examples{
MST(1)
}
