% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{p0}
\alias{p0}
\title{The stationary probabilities of the p0}
\usage{
p0(m = c(0.2, 0.3))
}
\arguments{
\item{m}{distribution parameters vector of sojourn times
in alternating environment states}
}
\description{
The stationary probabilities of the p0
}
\examples{
p0()
}
