% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{pi}
\alias{pi}
\title{Stationary probabilities for continuous time environment's state}
\usage{
pi(m = c(0.2, 0.3))
}
\arguments{
\item{m}{distribution parameters vector of sojourn times
in alternating environment states}
}
\value{
double
}
\description{
Stationary probabilities for continuous time environment's state
}
\examples{
pi()
}
