% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{probabilityOfNArrival}
\alias{probabilityOfNArrival}
\title{Probability of n arrival during time t jointly with final state j if initial state is i}
\usage{
probabilityOfNArrival(i, j, n, t, m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{n}{number of arrivals}

\item{t}{upper integration limit}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
double
}
\description{
Probability of n arrival during time t jointly with final state j if initial state is i
}
\examples{
probabilityOfNArrival(1, 0, 10, 3, m=c(0.5, 0.3), lambda=c(2, 1))
}
