% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{probabilityOfNArrivalW}
\alias{probabilityOfNArrivalW}
\title{Probability of n arrival during time t (without joint probability of j)}
\usage{
probabilityOfNArrivalW(i, n, t, m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{i}{MC i-th state}

\item{n}{number of arrivals}

\item{t}{time value}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
double
}
\description{
Probability of n arrival during time t (without joint probability of j)
}
\examples{
probabilityOfNArrivalW(1, 2, 3)
}
