% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{PrTr}
\alias{PrTr}
\title{Probability to have state j in the ending of the idle period, if initially we have state i}
\usage{
PrTr(i, j, m = c(0.2, 0.3), lambda = c(1, 2), tmax = 12)
}
\arguments{
\item{i}{MC i-th state}

\item{j}{MC j-th state}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}

\item{tmax}{upper integration limit}
}
\value{
double
}
\description{
Probability to have state j in the ending of the idle period, if initially we have state i
}
\examples{
PrTr(1, 0)
}
