% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{serviceDistribution}
\alias{serviceDistribution}
\title{Service distribution function}
\usage{
serviceDistribution(t)
}
\arguments{
\item{t}{time value}
}
\value{
service function value (double) given \code{t}
}
\description{
Service distribution function
}
