% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{stationaryProbabilitiesOfEmptyStates}
\alias{stationaryProbabilitiesOfEmptyStates}
\title{Stationary probabilities of the empty states in continuous time model}
\usage{
stationaryProbabilitiesOfEmptyStates(i, m = c(0.2, 0.3), lambda = c(1, 2))
}
\arguments{
\item{i}{MC i-th state}

\item{m}{distribution parameters vector of sojourn times
in alternating environment states}

\item{lambda}{Poisson flow intensity vector}
}
\value{
complex
}
\description{
Stationary probabilities of the empty states in continuous time model
}
