% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{p1}
\alias{p1}
\title{The stationary probabilities of the environment state 1}
\usage{
p1(m = c(0.2, 0.3))
}
\arguments{
\item{m}{distribution parameters vector of sojourn times
in alternating environment states}
}
\value{
stationary probability of the environment state 1 (double)
}
\description{
The stationary probabilities of the environment state 1
}
